got live notifications for pressure and gas working, updated readme

+++ software/firmware/include/adc.h
@@ -1,7 +1,7 @@
 #ifndef ADC_H
 #define ADC_H
 
-uint16_t getSample();
+uint32_t getSample();
 void adcConfig(uint8_t res, uint8_t pscale, uint8_t ref, uint8_t psel, 
                                                         uint8_t extref);
